+++ sdk/react-native/test/factory/didFactory.ts
@@ -1,7 +1,7 @@
 const cryptoRandomString = require('crypto-random-string')
 
 import { AffinidiWallet } from '../../'
-import { KeysService, DidDocumentService } from '@affinidi/common'
+import { KeysService, DidDocumentService, DidResolver } from '@affinidi/common'
 
 interface TestDid {
   seed: string
@@ -26,6 +26,9 @@ export const generateTestDIDs = async (): Promise<{
 }> => {
   let keysService
   let didDocumentService
+  const didResolverMock: DidResolver = {
+    resolveDid: () => Promise.resolve({ id: 'did:elem:ushJhdunHuhsecb_hscudTYj2h2e' }),
+  } as any
   const password = cryptoRandomString({ length: 32, type: 'ascii-printable' })
 
   const joloSeed = await AffinidiWallet.generateSeed('jolo')
@@ -36,8 +39,8 @@ export const generateTestDIDs = async (): Promise<{
 
   keysService = new KeysService(joloEncryptedSeed, password)
 
-  didDocumentService = new DidDocumentService(keysService)
-  const joloDidDocument = await didDocumentService.buildDidDocument()
+  didDocumentService = DidDocumentService.createDidDocumentService(keysService)
+  const joloDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
   const joloDid = joloDidDocument.id
 
   const joloPublicKey = KeysService.getPublicKey(joloSeedHex, 'jolo').toString('hex')
@@ -51,8 +54,8 @@ export const generateTestDIDs = async (): Promise<{
 
   keysService = new KeysService(elemEncryptedSeed, password)
 
-  didDocumentService = new DidDocumentService(keysService)
-  const elemDidDocument = await didDocumentService.buildDidDocument()
+  didDocumentService = DidDocumentService.createDidDocumentService(keysService)
+  const elemDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
   const elemDid = await didDocumentService.getMyDid()
 
   const elemPublicKey = KeysService.getPublicKey(elemSeedHex, 'elem').toString('hex')
@@ -69,8 +72,8 @@ export const generateTestDIDs = async (): Promise<{
 
   keysService = new KeysService(elemAltEncryptedSeed, password)
 
-  didDocumentService = new DidDocumentService(keysService)
-  const elemAltDidDocument = await didDocumentService.buildDidDocument()
+  didDocumentService = DidDocumentService.createDidDocumentService(keysService)
+  const elemAltDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
   const elemAltDid = await didDocumentService.getMyDid()
 
   const elemAltPublicKey = KeysService.getPublicKey(elemSeedHex, 'elem').toString('hex')